+++ util/set/set_with_memory_usage_test.go
@@ -1,4 +1,4 @@
-package aggfuncs
+package set
 
 import (
 	"fmt"
